Fix no newline at end of line issue

+++ mmdet3d/ops/pointnet2_utils/src/cuda_utils.h
@@ -8,7 +8,7 @@
 #define DIVUP(m,n) ((m) / (n) + ((m) % (n) > 0))
 
 inline int opt_n_threads(int work_size) {
-    const int pow_2 = std::log(static_cast<double>(work_size)) / std::log(2.0);
+    const int pow_2 = static_cast<int>(std::log(work_size) / std::log(2.0));
 
     return max(min(1 << pow_2, TOTAL_THREADS), 1);
 }
